mcp-server/requirements.txt
# MCP Server Core
mcp>=1.0.0

# Web Crawling & Content Extraction
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.26.0
html2text>=2024.2.0

# YAML Processing
PyYAML>=6.0.1
ruamel.yaml>=0.18.0

# URL Parsing & Validation
validators>=0.22.0
urllib3>=2.1.0

# Language Detection
langdetect>=1.0.9

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.1

# Logging & Monitoring
structlog>=24.1.0
colorama>=0.4.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Type Hints
typing-extensions>=4.9.0

# CLI
click>=8.1.7
rich>=13.7.0

# Error Handling
tenacity>=8.2.0